pyproject.toml
[project]
name = "scales"
version = "0.1.0"
dependencies = [
    "numpy",
    "openpnm",
    "numpy-stl",
    "scikit-image",
    "tqdm",
    "pylibtiff",
    "joblib",
    "vtk",
]
requires-python = ">= 3.9"
authors = [
    {name = "David Rieder", email = "d.r.rieder@tue.nl"},
]
maintainers = [
    {name = "David Rieder", email = "d.r.rieder@tue.nl"},
]
description = "A collection of useful python scripts"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["pnm", "image", "analysis", "permeability"]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python"
]

[project.urls]
Homepage = "https://github.com/hpc-dave/Scales"
Repository = "https://github.com/hpc-dave/Scales"